Lancei os popups
Lancei os popups de confirmacao, cancelamento e de config e tambem resolvi alguns b.o de gestao e cliente
